% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BuildPreCalmatrix_module.R
\name{BuildPreCalmatrix}
\alias{BuildPreCalmatrix}
\title{Title BuildPreCalmatrix}
\usage{
BuildPreCalmatrix(FeaturepreCalfile = FeaturepreCalfile)
}
\arguments{
\item{FeaturepreCalfile}{The name of FeaturepreCalfile}
}
\value{
preCalmatrix
}
\description{
Title BuildPreCalmatrix
}
\examples{
BuildPreCalmatrix(FeaturepreCalfile)
}
