% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FeaturepreCalfile}
\alias{FeaturepreCalfile}
\title{FeaturepreCalfile}
\format{
A vector of file name
\describe{
}
}
\source{
<https://github.com/wangxlab/IndepthPathwayData>
}
\usage{
FeaturepreCalfile
}
\description{
FeaturepreCalfile
}
\keyword{datasets}
