% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{preCalmatrix}
\alias{preCalmatrix}
\title{preCalmatrix}
\format{
A matrix
\describe{
}
}
\source{
<https://github.com/wangxlab/IndepthPathwayData>
}
\usage{
preCalmatrix
}
\description{
preCalmatrix data
}
\keyword{datasets}
